requirements.txt
Django==2.0.2

django-braces==1.11.0
django-compressor==2.2
django-extensions==1.9.9
celery==4.1
django-celery
flower
anyjson
gunicorn==19.7.1
psycopg2==2.7.3.2
django-material
django-import-export
django-jet
